# Add optimized methods