Define `String` and `string` for `CategoricalString` since the `AbstractString` fallbacks are less efficient. Add `get` and `get!` methods for `CatValues` in a `CategoricalPool` to avoid a dict lookup when the pool is the same.

```julia
x = categorical(rand(["a", "b", "c", "d"], 10000));
f(x) = (y = x[1]; for i in eachindex(x); x[i] = y; end)
v = x[1]

# Before
julia> @btime f(x);
  506.536 μs (10000 allocations: 312.50 KiB)

julia> @btime String($v);
  50.742 ns (3 allocations: 160 bytes)

# After
julia> @btime f(x);
  13.527 μs (0 allocations: 0 bytes)

julia> @btime String($v);
  3.846 ns (0 allocations: 0 bytes)
```

Failures on nightly are unrelated and due to changes in dates printing.